Backout 368cd30cccda for bustage

+++ netwerk/test/TestIncrementalDownload.cpp
@@ -57,8 +57,7 @@ NS_IMETHODIMP
 FetchObserver::OnStopRequest(nsIRequest *request, nsISupports *context,
                              nsresult status)
 {
-  printf("FetchObserver::OnStopRequest [status=%x]\n",
-         static_cast<uint32_t>(status));
+  printf("FetchObserver::OnStopRequest [status=%x]\n", status);
 
   QuitPumpingEvents();
   return NS_OK;
@@ -122,8 +121,7 @@ main(int argc, char **argv)
 
   rv = DoIncrementalFetch(argv[1], argv[2], chunkSize, interval);
   if (NS_FAILED(rv))
-    fprintf(stderr, "ERROR: DoIncrementalFetch failed [%x]\n",
-            static_cast<uint32_t>(rv));
+    fprintf(stderr, "ERROR: DoIncrementalFetch failed [%x]\n", rv);
 
   NS_ShutdownXPCOM(nullptr);
   return 0;
