+++ netwerk/test/TestPageLoad.cpp
@@ -180,8 +180,7 @@ MyListener::OnDataAvailable(nsIRequest *req, nsISupports *ctxt,
     }
 
     if (NS_FAILED(rv)) {
-      printf(">>> stream->Read failed with rv=%x\n",
-             static_cast<uint32_t>(rv));
+      printf(">>> stream->Read failed with rv=%x\n", rv);
       return rv;
     }
